spawns an enemy immediately after creation

+++ Project_Digger/Objects/Spawner.hpp
@@ -13,6 +13,6 @@ public:
 
 private:
 
-	size_t lastSpawnTime;
+	int lastSpawnTime;
 
 };
